respond to pr feedback
