+++ .nvmrc
@@ -1,1 +1,1 @@
-12.14
+12.14.1
