+++ src/Apps/IdentityVerification/__tests__/Utils/IdentityVerificationAppTestPage.tsx
@@ -1,11 +1,12 @@
 import { RootTestPage } from "DevTools/RootTestPage"
 
-export class IdentityVerificationTestPage extends RootTestPage {
+export class IdentityVerificationAppTestPage extends RootTestPage {
   get startVerificationButton() {
     return this.find("button").filterWhere(btn =>
       btn.text().includes("Continue to verification")
     )
   }
+
   async clickStartVerification() {
     this.startVerificationButton.simulate("click")
     await this.update()
