+++ src/Apps/IdentityVerification/routes.tsx
@@ -5,12 +5,12 @@ import { graphql } from "react-relay"
 export const routes: RouteConfig[] = [
   {
     path: "/identity-verification/:id",
-    getComponent: () => loadable(() => import("./IdentityVerificationPage")),
+    getComponent: () => loadable(() => import("./IdentityVerificationApp")),
     query: graphql`
-      query routes_IdentityVerificationPageQuery($id: String!)
+      query routes_IdentityVerificationAppQuery($id: String!)
         @raw_response_type {
         me {
-          ...IdentityVerificationPage_me @arguments(id: $id)
+          ...IdentityVerificationApp_me @arguments(id: $id)
         }
       }
     `,
