+++ src/Apps/getAppRoutes.tsx
@@ -3,6 +3,7 @@ import { routes as artworkRoutes } from "Apps/Artwork/routes"
 import { routes as auctionRoutes } from "Apps/Auction/routes"
 import { collectRoutes } from "Apps/Collect2/collectRoutes"
 import { conversationRoutes } from "Apps/Conversation/routes"
+import { routes as identityVerificationRoutes } from "Apps/IdentityVerification/routes"
 import { routes as orderRoutes } from "Apps/Order/routes"
 import { routes as searchRoutes } from "Apps/Search/routes"
 import { makeAppRoutes } from "Artsy/Router/makeAppRoutes"
@@ -30,6 +31,9 @@ export function getAppRoutes(): RouteConfig[] {
     },
     {
       routes: searchRoutes,
+    },
+    {
+      routes: identityVerificationRoutes,
     },
   ])
 }
